Document more of bootstrap and configure
There are a few details that were mentioned in HACKING that had
not made there way over, and seem useful. It's possible that this
part of the documentation is now getting a bit fiddly, and we may
want to split "first time just get there" from "rest of the stuff",
but for now this is better than not having the details available.